+++ docs/conf.py
@@ -2,14 +2,8 @@
 # pylint: disable=invalid-name
 import os
 import time
-from pathlib import Path
-from typing import List
 from urllib.parse import urlparse
 
-from docutils import nodes
-from sphinx.application import Sphinx
-from sphinx.domains.index import IndexDirective
-from sphinx.errors import SphinxError
 
 GIT_BRANCH = os.environ.get("SPHINX_GITHUB_BRANCH", "") or os.environ.get("GITHUB_REF", "").split("/", 2)[-1] or "main"
 GIT_URL = "https://github.com/Robpol86/robpol86.com"
@@ -22,6 +16,7 @@ exclude_patterns = []
 extensions = [
     "myst_parser",  # https://myst-parser.readthedocs.io
     "notfound.extension",  # https://sphinx-notfound-page.readthedocs.io
+    "robpol86_com.legacy",
     "sphinx_copybutton",  # https://sphinx-copybutton.readthedocs.io
     "sphinx_disqus.disqus",  # https://sphinx-disqus.readthedocs.io
     "sphinx_external_toc",  # https://sphinx-external-toc.readthedocs.io
@@ -118,53 +113,3 @@ ogp_site_url = html_baseurl
 ogp_type = "website"
 ogp_use_first_image = True
 sitemap_url_scheme = "{link}"
-
-
-def render_robots_txt(app: Sphinx, _):
-    """Parse Jinja2 templating in robots.txt file.
-
-    :param app: Sphinx application object.
-    :param _: Unused.
-    """
-    robots_txt_path = Path(app.outdir) / "robots.txt"
-    if robots_txt_path.is_file():
-        contents = robots_txt_path.read_text(encoding="utf8")
-        context = dict(app.config.html_context, config=app.config)
-        rendered = app.builder.templates.render_string(contents, context)
-        robots_txt_path.write_text(rendered, encoding="utf8")
-
-
-class TagsDirective(IndexDirective):
-    """Enhanced Sphinx index directive so it acts more like a tag manager."""
-
-    def run(self) -> List:
-        """Called by Sphinx."""
-        index_node, target_node = super().run()
-        tags = [t[1] for t in index_node["entries"]]
-        if not tags:
-            return [index_node, target_node]
-        if tags != sorted(tags):
-            raise SphinxError(f"Tags not in alphabetical order in document {self.env.docname}")
-
-        # Build nodes.
-        human_readable_tag_list = nodes.emphasis("Tags: ", "Tags: ")
-        idx_last = len(tags) - 1
-        for idx, tag in enumerate(tags):
-            tag_node = nodes.inline(tag, tag, classes=["guilabel"])
-            uri = f"{html_baseurl}genindex.html#{tag[0].upper()}"
-            linked_tag_node = nodes.reference("", "", tag_node, refuri=uri, internal=True)
-            # Insert.
-            human_readable_tag_list.append(linked_tag_node)
-            if idx != idx_last:
-                human_readable_tag_list.append(nodes.Text(", ", ", "))
-
-        return [index_node, target_node, nodes.paragraph("", "", human_readable_tag_list)]
-
-
-def setup(app: Sphinx):
-    """Called by Sphinx.
-
-    :param app: Sphinx application object.
-    """
-    app.connect("build-finished", render_robots_txt)
-    app.add_directive("tags", TagsDirective)
